requirements.txt
Flask==3.0.2
python-barcode==0.15.1
Pillow==10.2.0
gunicorn==21.2.0 